// types/styles.types.ts
interface StylesProps {
    container: object;
    title: object;
    inputContainer: object;
    input: object;
    addButton: object;
    addButtonText: object;
    task: object;
    taskTextContainer: object;
    taskText: object;
    completedTask: object;
    deleteButton: object;
    deleteButtonText: object;
    prioritySelector: object;
    priorityButton: object;
    activePriority: object;
    priorityButtonText: object;
    high: object;
    medium: object;
    low: object;
    [key: string]: any;
  }
  
  export default StylesProps;